feat: write wrapper for boost python
issue #461
